+++ sktime/classification/early_classification/base/_base.py
